models.py razstavljen v več datotek
