+++ nadlogar/problems/models/__init__.py
@@ -1,0 +1,4 @@
+from .razno import *
+from .razstavljanje import *
+from .mnozice import *
+from .veccleniki import *
